GV-623 added typescript in the component of PasswordChange
